control for light

+++ src/shaders/water.vertex.glsl
@@ -4,8 +4,8 @@ attribute vec2 textureCoord;
 uniform mat4 model;
 uniform mat4 projection;
 
-varying lowp vec2 vTextureCoord;
-varying lowp vec4 clipSpace;
+varying highp vec2 vTextureCoord;
+varying highp vec4 clipSpace;
 
 void main() {
   clipSpace = projection * model * position;
